refactor: reconstruct control component to live componenet
