+++ src/domains/control/hooks/solving/question/useControlSolvingQuestion.ts
@@ -4,7 +4,6 @@ import type { QuestionResponseType } from "@/app.constants";
 import { notify } from "@/components/Toast";
 import useControlQuestionSubmitAnswerStore from "@/domains/control/stores/question/useControlQuestionSubmitAnswerStore";
 import { apiHooks } from "@/libs/api";
-import type { UpdateQuestionStatusApiRequest } from "@/libs/types";
 
 //
 //
@@ -22,12 +21,7 @@ export interface UseControlSolvingQuestionReturn<TData> {
 	question?: TData;
 	refetchQuestion: () => void;
 	submitAnswer: () => Promise<boolean>;
-	handleAccessOpen: () => void;
-	handleAccessClose: () => void;
-	handleSolveOpen: () => void;
-	handleSolveClose: () => void;
 	isLoading: boolean;
-	isStatusUpdating: boolean;
 }
 
 //
@@ -77,19 +71,7 @@ const useControlSolvingQuestion = <
 		},
 	);
 
-	const questionSet = questionSetData?.data;
 	const question = questionData?.data;
-
-	const { mutate: updateStatus, isPending: isStatusUpdating } =
-		apiHooks.useMutation(
-			"patch",
-			"/api/v1/question-sets/{questionSetId}/questions/{questionId}/status",
-			{
-				onSuccess: () => {
-					refetch();
-				},
-			},
-		);
 
 	const { mutateAsync: postAnswer } = apiHooks.useMutation(
 		"post",
@@ -155,69 +137,6 @@ const useControlSolvingQuestion = <
 		return res.isSuccess ?? false;
 	};
 
-	/**
-	 *
-	 */
-	const handleStatusUpdate = (
-		status: UpdateQuestionStatusApiRequest["statusType"],
-	) => {
-		if (isStatusUpdating) {
-			return;
-		}
-
-		if (questionSet?.status === "BEFORE") {
-			notify.warn("문제 시작 후 문제 공개 및 제출 허용이 가능합니다.");
-
-			return;
-		}
-
-		updateStatus({
-			params: {
-				path: {
-					questionSetId,
-					questionId,
-				},
-			},
-			body: {
-				statusType: status,
-			},
-		});
-	};
-
-	/**
-	 *
-	 */
-	const handleAccessOpen = () => {
-		handleStatusUpdate("ACCESS_PERMISSION");
-	};
-
-	/**
-	 *
-	 */
-	const handleAccessClose = () => {
-		handleStatusUpdate("NOT_OPEN");
-	};
-
-	/**
-	 *
-	 */
-	const handleSolveOpen = () => {
-		if (question?.questionStatusType === "NOT_OPEN") {
-			notify.warn("문제 시작 후 문제 공개 및 제출 허용이 가능합니다.");
-
-			return;
-		}
-
-		handleStatusUpdate("SOLVE_PERMISSION");
-	};
-
-	/**
-	 *
-	 */
-	const handleSolveClose = () => {
-		handleStatusUpdate("ACCESS_PERMISSION");
-	};
-
 	//
 	//
 	// biome-ignore lint/correctness/useExhaustiveDependencies: fetch on questionId change
@@ -230,12 +149,7 @@ const useControlSolvingQuestion = <
 		question: question as TData | undefined,
 		refetchQuestion: () => refetch(),
 		submitAnswer,
-		handleAccessOpen,
-		handleAccessClose,
-		handleSolveOpen,
-		handleSolveClose,
 		isLoading: isQuestionLoading || isQuestionSetLoading,
-		isStatusUpdating,
 	};
 };
