+++ src/domains/control/pages/common/solving/question/ControlSolvingQuestionFillBlank.tsx
