+++ src/domains/control/pages/common/solving/question/ControlSolvingQuestionPanel.tsx
@@ -2,12 +2,12 @@ import clsx from "clsx";
 import { Check, Pencil, X } from "lucide-react";
 import { useEffect, useState } from "react";
 import { useParams } from "react-router-dom";
+import type { ReactNode } from "react";
 import Button from "@/components/Button";
-import { Switch } from "@/components/switch/Switch";
 import Tooltip from "@/components/Tooltip";
 import ControlSolvingQuestionContent from "@/domains/control/components/solving/question/ControlSolvingQuestionContent";
 import useControlSolvingQuestion from "@/domains/control/hooks/solving/question/useControlSolvingQuestion";
-import type { QuestionApiResponse, QuestionType } from "@/libs/types";
+import type { QuestionType } from "@/libs/types";
 import ControlSolvingQuestionFillBlank from "./ControlSolvingQuestionFillBlank";
 import ControlSolvingQuestionMultiple from "./ControlSolvingQuestionMultiple";
 import ControlSolvingQuestionOrdering from "./ControlSolvingQuestionOrdering";
@@ -17,25 +17,27 @@ import ControlSolvingQuestionShort from "./ControlSolvingQuestionShort";
 //
 //
 
-const ControlSolvingQuestion = () => {
+interface ControlSolvingQuestionPanelProps {
+	topControls?: ReactNode;
+}
+
+//
+//
+//
+
+const ControlSolvingQuestionPanel = ({
+	topControls,
+}: ControlSolvingQuestionPanelProps) => {
 	const [isEditing, setIsEditing] = useState(false);
 	const [isMouseOverOnEditButton, setIsMouseOverOnEditButton] = useState(false);
-	const [updateStatusType, setUpdateStatusType] = useState<
-		"ACCESS" | "SOLVE" | null
-	>(null);
 
 	const questionSetId = Number(useParams().questionSetId);
 	const questionId = Number(useParams().questionId);
 
 	const {
 		hasSubmitAnswerPayload,
-		isStatusUpdating,
 		question,
 		refetchQuestion,
-		handleAccessOpen,
-		handleAccessClose,
-		handleSolveOpen,
-		handleSolveClose,
 		submitAnswer,
 	} = useControlSolvingQuestion({
 		questionSetId,
@@ -66,76 +68,6 @@ const ControlSolvingQuestion = () => {
 		if (res) {
 			setIsEditing(false);
 		}
-	};
-
-	/**
-	 *
-	 */
-	const handleAccessSwitchChange = (checked: boolean) => {
-		setUpdateStatusType("ACCESS");
-
-		if (checked) {
-			handleAccessOpen();
-		} else {
-			handleAccessClose();
-		}
-	};
-
-	/**
-	 *
-	 */
-	const handleSolveSwitchChange = (checked: boolean) => {
-		setUpdateStatusType("SOLVE");
-
-		if (checked) {
-			handleSolveOpen();
-		} else {
-			handleSolveClose();
-		}
-	};
-
-	/**
-	 *
-	 */
-	const renderQuestionControlButtons = () => {
-		const allowedAccessTypes: QuestionApiResponse["questionStatusType"][] = [
-			"ACCESS_PERMISSION",
-			"SOLVE_PERMISSION",
-		];
-		const allowedSolveType: QuestionApiResponse["questionStatusType"][] = [
-			"SOLVE_PERMISSION",
-		];
-
-		const isSolveSwitchLoading =
-			updateStatusType === "SOLVE" &&
-			!allowedSolveType.includes(question?.questionStatusType) &&
-			isStatusUpdating;
-
-		return (
-			<div className="flex gap-gap-9">
-				<Switch.Root
-					checked={allowedAccessTypes.includes(question?.questionStatusType)}
-					onChange={handleAccessSwitchChange}
-				>
-					<Switch.Label>문제 공개</Switch.Label>
-					<Switch.Toggle />
-				</Switch.Root>
-				<Switch.Root
-					checked={allowedSolveType.includes(question?.questionStatusType)}
-					loading={isSolveSwitchLoading}
-					onChange={handleSolveSwitchChange}
-				>
-					<Switch.Label>제출 허용</Switch.Label>
-					<Tooltip
-						open={isSolveSwitchLoading}
-						message="제출 허용은 5초 이내에 활성화됩니다."
-						variant="primary"
-					>
-						<Switch.Toggle />
-					</Tooltip>
-				</Switch.Root>
-			</div>
-		);
 	};
 
 	/**
@@ -201,18 +133,6 @@ const ControlSolvingQuestion = () => {
 	/**
 	 *
 	 */
-	const renderQuestionContent = () => {
-		return (
-			<ControlSolvingQuestionContent
-				content={question?.content}
-				imgUrl={question?.imageUrl}
-			/>
-		);
-	};
-
-	/**
-	 *
-	 */
 	const renderQuestionAnswer = () => {
 		switch (question?.type as QuestionType) {
 			case "MULTIPLE":
@@ -235,26 +155,20 @@ const ControlSolvingQuestion = () => {
 		setIsEditing(false);
 	}, [questionId]);
 
-	//
-	//
-	//
-	useEffect(() => {
-		if (!isStatusUpdating) {
-			setUpdateStatusType(null);
-		}
-	}, [isStatusUpdating]);
-
 	return (
 		<div className="flex flex-col gap-gap-11 min-w-0">
 			<div className="flex justify-between">
-				{isEditing ? renderCancelButton() : renderQuestionControlButtons()}
+				{isEditing ? renderCancelButton() : (topControls ?? <div />)}
 				{renderEditButton()}
 			</div>
 
-			{renderQuestionContent()}
+			<ControlSolvingQuestionContent
+				content={question?.content}
+				imgUrl={question?.imageUrl}
+			/>
 			{renderQuestionAnswer()}
 		</div>
 	);
 };
 
-export default ControlSolvingQuestion;
+export default ControlSolvingQuestionPanel;
