+++ src/domains/control/pages/common/solving/question/ControlSolvingQuestionShort.tsx
