+++ src/domains/control/pages/common/solving/submission/ControlSolvingSubmissionTableBody.tsx
