+++ src/domains/control/pages/common/solving/submission/ControlSolvingSubmissionTabs.tsx
