+++ src/domains/control/pages/live/participant/ControlLiveParticipantScorerRanking.tsx
@@ -1,22 +1,22 @@
-import { Award } from "lucide-react";
+import { Coins } from "lucide-react";
 import { useParams } from "react-router-dom";
 import { Table } from "@/components/table";
 import type { UserApiResponse } from "@/libs/types";
-import ControlParticipantRankingUser from "../../components/participant/ControlParticipantRankingUser";
-import { ControlParticipantRankingPanel } from "../../components/participant/ranking-panel";
-import useControlParticipantRanking from "../../hooks/paticipant/useControlParticipantRanking";
+import ControlParticipantRankingUser from "../../../components/participant/ControlParticipantRankingUser";
+import { ControlParticipantRankingPanel } from "../../../components/participant/ranking-panel";
+import useControlParticipantRanking from "../../../hooks/paticipant/useControlParticipantRanking";
 
 //
 //
 //
 
-const ControlParticipantCorrectRanking = () => {
+const ControlLiveParticipantScorerRanking = () => {
 	const questionSetId = Number(useParams().questionSetId);
 
 	const {
 		ranking,
-		handleRankingRowParticipantsChange,
 		eliminatedParticipants,
+		handleRankingRowParticipantsChange,
 		checkIsAllUsersActive,
 		selectedRank,
 		handleSelectRank,
@@ -25,7 +25,7 @@ const ControlParticipantCorrectRanking = () => {
 		handleEliminateAllParticipants,
 	} = useControlParticipantRanking({
 		questionSetId,
-		type: "CORRECT" as const,
+		type: "SCORER" as const,
 	});
 
 	/**
@@ -37,20 +37,6 @@ const ControlParticipantCorrectRanking = () => {
 		} else {
 			handleEliminateAllParticipants();
 		}
-	};
-
-	/**
-	 *
-	 */
-	const renderRankCell = (rank: number, count: number) => {
-		return (
-			<span className="flex items-center gap-gap-5">
-				<span>{rank}등</span>
-				<span className="px-padding-5 py-padding-1 bg-color-primary-10 rounded-full typo-body-xsmall-bold text-color-primary-50">
-					{count}명
-				</span>
-			</span>
-		);
 	};
 
 	/**
@@ -76,11 +62,11 @@ const ControlParticipantCorrectRanking = () => {
 	return (
 		<ControlParticipantRankingPanel.Root>
 			<ControlParticipantRankingPanel.Header
-				icon={<Award />}
-				title="정답수 기준 등수"
+				icon={<Coins />}
+				title="선착순 기준 등수"
 			/>
 			<ControlParticipantRankingPanel.Selector
-				ranking="정답자"
+				ranking="선착순"
 				selectedRank={selectedRank}
 				onRankChange={handleSelectRank}
 				onApplySelection={handleApplyRankSelection}
@@ -98,7 +84,7 @@ const ControlParticipantCorrectRanking = () => {
 								// biome-ignore lint/suspicious/noArrayIndexKey: ranking order is stable and only user data changes
 								key={index}
 								checked={checkIsAllUsersActive(users)}
-								rankCell={renderRankCell(index + 1, users?.length ?? 0)}
+								rankCell={`${index + 1}등`}
 								nameCell={renderNameCell(users)}
 								onChange={(checked) =>
 									handleRankingRowParticipantsChange(checked, users)
@@ -113,4 +99,4 @@ const ControlParticipantCorrectRanking = () => {
 	);
 };
 
-export default ControlParticipantCorrectRanking;
+export default ControlLiveParticipantScorerRanking;
